+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added an option to always show the module target, even for error, warning, and
+  info messages. This is useful for debugging and for setting up module filters.
+
 ## [0.2.0] - 2023-03-01
 
 ### Added